Pull out reconcileCMRMetadata tests.
This separation just makes sense to me.
